pyproject.toml
[tool.poetry]
name = "text2signal"
version = "0.1.0"
description = ""
authors = ["Process AI team <DL_63BDEE2247BB9AFD5833B18E@global.corp.sap>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
python-dotenv = "1.0.0"
deepdiff = "6.7.1"
pyyaml = "6.0.1"
sap-doxml-commons = { extras = ["data"], version = "4.47.0", source = "internal-hyperspace-deploy" }
sap-llm-eval-toolkit = { version = "0.1.11", source = "internal-hyperspace" }
signavio = {version = "^2.0.0", source = "signavio"}

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
jupyter = "^1.0.0"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
sqlglot = "^20.11.0"
transformers = "^4.37.2"
sentence-transformers = "^2.3.1"
#openai = "^1.11.1"
openai = "^0.28.1"
tiktoken = "^0.5.2"

[[tool.poetry.source]]
name = "hyperspace"
url = "https://common.repositories.cloud.sap/artifactory/api/pypi/deploy-releases-hyperspace-pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "internal-hyperspace"
url = "https://int.repositories.cloud.sap/artifactory/api/pypi/proxy-deploy-releases-hyperspace-pypi/simple/"
#      https://int.repositories.cloud.sap/artifactory/api/pypi/proxy-deploy-releases-hyperspace-pypi/simple
priority = "primary"


[[tool.poetry.source]]
name = "internal-hyperspace-deploy"
url = "https://int.repositories.cloud.sap/artifactory/api/pypi/deploy-releases-pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "default"


[[tool.poetry.source]]
name = "signavio"
url = "https://common.repositories.cloud.sap/artifactory/api/pypi/signavio-pypi/simple"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "W", # pycodestyle warnings
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "TCH", # flake8-type-checking
    "PERF", # Perflint
    "C4", # flake8-comprehensions
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "D" # pydocstyle
]
ignore = [
    "A003", # class attribute shadowing builtin
    "E501" # line too long, handled by black
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

target-version = "py311"
line-length = 120

[tool.ruff.isort]
known-first-party = ["app"]

[tool.ruff.per-file-ignores]
"tests/*" = ["D"]
"__init__.py" = ["D"]
"run_api_locally.py" = ["D"]

[tool.ruff.pydocstyle]
convention = "pep257"

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
